+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [2.3.0]
 
+### Added
+
+- First implementation (preview) of automatic registration of plugin steps via reflection - https://github.com/DynamicsValue/fake-xrm-easy/issues/11
+
+
 ### Changed
 
 - IXrmFakedContextPluginExtensions is now IXrmBaseContextPluginExtensions. All extensions methods were changed to use the IXrmBaseContext interface as opposed to the IXrmFakedContext interface only so they could be used by both XrmFakedContext and XrmRealContext classes. - https://github.com/DynamicsValue/fake-xrm-easy/issues/35